Compile BSP and bootloader as C++
