--- boot/crc.c
@@ -1,67 +1,0 @@
-/**************************************************************************//**
- * @file
- * @brief CRC calculation routines
- * @author Energy Micro AS
-* @version 1.63
- ******************************************************************************
- * @section License
- * <b>(C) Copyright 2009 Energy Micro AS, http://www.energymicro.com</b>
- ******************************************************************************
- *
- * This source code is the property of Energy Micro AS. The source and compiled
- * code may only be used on Energy Micro "EFM32" microcontrollers.
- *
- * This copyright notice may not be removed from the source code nor changed.
- *
- * DISCLAIMER OF WARRANTY/LIMITATION OF REMEDIES: Energy Micro AS has no
- * obligation to support this Software. Energy Micro AS is providing the
- * Software "AS IS", with no express or implied warranties of any kind,
- * including, but not limited to, any implied warranties of merchantability
- * or fitness for any particular purpose or warranties against infringement
- * of any proprietary rights of a third party.
- *
- * Energy Micro AS will not be liable for any consequential, incidental, or
- * special damages, or any other relief, or for any claim by any third party,
- * arising from your use of this Software.
- *
- *****************************************************************************/
-#include "crc.h"
-#include "em_system.h"
-
-/**************************************************************************//**
- * @brief
- *   This function calculates the CRC-16-CCIT checksum of a memory range.
- *  
- * @note
- *   This implementation uses an initial value of 0, while some implementations
- *   of CRC-16-CCIT uses an initial value of 0xFFFF. If you wish to
- *   precalculate the CRC before uploading the binary to the bootloader you 
- *   can use this function. However, keep in mind that the 'v' and 'c' commands
- *   computes the crc of the entire flash, so any bytes not used by your
- *   application will have the value 0xFF.
- *
- * @param start
- *   Pointer to the start of the memory block
- *
- * @param end
- *   Pointer to the end of the block. This byte is not included in the computed
- *   CRC.
- *
- * @return
- *   The computed CRC value.
- *****************************************************************************/
-uint16_t CRC_calc(uint8_t *start, uint8_t *end)
-{
-  uint16_t crc = 0x0;
-  uint8_t  *data;
-
-  for (data = start; data < end; data++)
-  {
-    crc  = (crc >> 8) | (crc << 8);
-    crc ^= *data;
-    crc ^= (crc & 0xff) >> 4;
-    crc ^= crc << 12;
-    crc ^= (crc & 0xff) << 5;
-  }
-  return crc;
-}
